+++ packages/react-app/src/helpers/index.js
@@ -1,1 +1,2 @@
 export { default as Transactor } from "./Transactor";
+export { default as logoutOfWeb3Modal } from "./logoutOfWeb3Modal";
