+++ packages/react-app/src/hooks/index.js
@@ -1,9 +1,5 @@
-export { default as useBalance } from "./Balance";
 export { default as useContractLoader } from "./ContractLoader";
 export { default as useExchangePrice } from "./ExchangePrice";
 export { default as useGasPrice } from "./GasPrice";
-export { default as useLookupAddress } from "./LookupAddress";
-export { default as useOnBlock } from "./OnBlock";
-export { default as usePoller } from "./Poller";
 export { default as useUserSigner } from "./UserSigner";
 export { default as useUpdateParcels } from "./UpdateParcels";
